Large rewrite of ScanChannel
I created a test file with a fake discord.Channel class to test the
implementation of the ScanChannel class as I could not properly figure
out exactly how to do this.

Using that knowledge I rebuilt this class from scratch. It is now working
properly when testing in my personal Discord.